--- a/docs/production-optimization-phase5-plan.md
+++ b/docs/production-optimization-phase5-plan.md
@@ -1,0 +1,246 @@
+# Production Optimization Phase 5 - Implementation Plan
+
+## Overview
+Phase 5 focuses on comprehensive production optimization including performance improvements, scalability enhancements, and robust monitoring systems to ensure v1z3r can handle enterprise-level workloads.
+
+## Current Performance Analysis
+
+### Identified Bottlenecks
+1. **Bundle Size**: Large initial load due to Three.js and WebGL libraries
+2. **WebGL Performance**: Potential frame drops during complex effects
+3. **Audio Processing**: Real-time FFT calculations could be optimized
+4. **Memory Usage**: Potential leaks in long-running sessions
+5. **Network Latency**: WebSocket reconnection and state sync delays
+6. **Database Queries**: DynamoDB scan operations could be optimized
+
+### Performance Metrics (Baseline)
+```
+Current Performance:
+- Initial Bundle Size: ~2.5MB (uncompressed)
+- Time to Interactive: ~3.5s
+- WebGL FPS: 45-60fps (varies by device)
+- Memory Usage: 150-300MB (grows over time)
+- WebSocket Latency: 50-200ms
+- Cold Start (Lambda): 800-1200ms
+```
+
+## Optimization Strategy
+
+### 1. Performance Monitoring & Metrics
+- Real User Monitoring (RUM) enhancement
+- Core Web Vitals tracking
+- Custom performance metrics
+- Error tracking and alerting
+- Resource usage monitoring
+
+### 2. Bundle Optimization
+- Advanced code splitting
+- Tree shaking optimization
+- Dynamic imports for heavy libraries
+- Service Worker caching
+- Compression and minification
+
+### 3. Scalability Architecture
+- Auto-scaling infrastructure
+- Load balancing strategies
+- Database connection pooling
+- Caching layers (Redis/ElastiCache)
+- CDN optimization
+
+### 4. Runtime Performance
+- WebGL shader optimization
+- Audio processing optimization
+- Memory leak prevention
+- Efficient state management
+- Background task optimization
+
+### 5. Monitoring & Alerting
+- CloudWatch dashboards
+- Application insights
+- Error tracking (Sentry integration)
+- Performance budgets
+- Health check endpoints
+
+## Implementation Plan
+
+### Week 1: Performance Monitoring Foundation
+1. **Enhanced RUM Implementation**
+   - Core Web Vitals tracking
+   - Custom performance metrics
+   - User experience analytics
+   - Real-time performance dashboards
+
+2. **Monitoring Infrastructure**
+   - CloudWatch custom metrics
+   - Application performance monitoring
+   - Error tracking and alerting
+   - Performance budget enforcement
+
+### Week 2: Bundle & Runtime Optimization
+1. **Bundle Size Reduction**
+   - Advanced code splitting
+   - Tree shaking optimization
+   - Dynamic imports for heavy modules
+   - Webpack bundle analyzer integration
+
+2. **Runtime Performance**
+   - WebGL optimization
+   - Audio processing improvements
+   - Memory management enhancements
+   - State synchronization optimization
+
+### Week 3: Scalability & Infrastructure
+1. **Scalability Enhancements**
+   - Auto-scaling configuration
+   - Load balancing optimization
+   - Database performance tuning
+   - Caching layer implementation
+
+2. **Production Infrastructure**
+   - CDN optimization
+   - Static asset compression
+   - Connection pooling
+   - Health monitoring
+
+### Week 4: Testing & Validation
+1. **Performance Testing**
+   - Load testing with realistic scenarios
+   - Stress testing for peak loads
+   - Performance regression testing
+   - Mobile performance validation
+
+2. **Production Deployment**
+   - Blue-green deployment setup
+   - Canary release process
+   - Rollback procedures
+   - Production validation
+
+## Technical Implementation
+
+### Performance Monitoring Stack
+```typescript
+// Enhanced RUM with Core Web Vitals
+interface PerformanceMetrics {
+  // Core Web Vitals
+  LCP: number;  // Largest Contentful Paint
+  FID: number;  // First Input Delay
+  CLS: number;  // Cumulative Layout Shift
+  FCP: number;  // First Contentful Paint
+  TTFB: number; // Time to First Byte
+  
+  // Custom Metrics
+  webglInitTime: number;
+  audioContextDelay: number;
+  effectSwitchTime: number;
+  memoryUsage: number;
+  frameRate: number;
+}
+```
+
+### Bundle Optimization Strategy
+```javascript
+// Webpack optimization config
+module.exports = {
+  optimization: {
+    splitChunks: {
+      chunks: 'all',
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          name: 'vendors',
+          priority: 10,
+          reuseExistingChunk: true,
+        },
+        threejs: {
+          test: /[\\/]node_modules[\\/]three[\\/]/,
+          name: 'threejs',
+          priority: 20,
+        },
+        webgl: {
+          test: /[\\/]src[\\/].*webgl.*[\\/]/,
+          name: 'webgl',
+          priority: 15,
+        },
+      },
+    },
+    usedExports: true,
+    sideEffects: false,
+  },
+};
+```
+
+### Scalability Architecture
+```yaml
+# Auto-scaling configuration
+AutoScaling:
+  MinSize: 2
+  MaxSize: 20
+  TargetCPU: 70%
+  TargetMemory: 80%
+  ScaleOutCooldown: 300s
+  ScaleInCooldown: 600s
+  
+LoadBalancer:
+  HealthCheck:
+    Path: /api/health
+    Interval: 30s
+    Timeout: 5s
+    HealthyThreshold: 2
+    UnhealthyThreshold: 3
+```
+
+## Success Metrics
+
+### Performance Targets
+- **Bundle Size**: < 2MB (20% reduction)
+- **Time to Interactive**: < 2.5s (30% improvement)
+- **WebGL FPS**: 60fps consistently
+- **Memory Usage**: < 200MB stable
+- **WebSocket Latency**: < 100ms
+- **Cold Start**: < 500ms
+
+### Scalability Targets
+- **Concurrent Users**: 1000+ simultaneous
+- **Request Throughput**: 10,000 req/min
+- **Database Performance**: < 100ms query time
+- **Availability**: 99.9% uptime
+- **Auto-scaling**: < 60s response time
+
+### Monitoring Targets
+- **Error Rate**: < 0.1%
+- **Alert Response**: < 5min
+- **Performance Budget**: 100% compliance
+- **Core Web Vitals**: All "Good" ratings
+- **User Satisfaction**: > 95%
+
+## Risk Assessment
+
+### Technical Risks
+- **Performance Regression**: Mitigated by comprehensive testing
+- **Scalability Issues**: Addressed by gradual load increase
+- **Breaking Changes**: Prevented by feature flags
+- **Memory Leaks**: Monitored by automated testing
+
+### Mitigation Strategies
+- Feature flags for gradual rollout
+- Comprehensive performance testing
+- Automated monitoring and alerting
+- Rollback procedures for quick recovery
+
+## Dependencies
+- CloudWatch for monitoring
+- ElastiCache for caching
+- Application Load Balancer
+- Auto Scaling Groups
+- Performance testing tools
+
+## Deliverables
+1. Enhanced performance monitoring system
+2. Optimized bundle and runtime performance
+3. Scalable infrastructure configuration
+4. Comprehensive monitoring and alerting
+5. Production deployment pipeline
+6. Performance testing suite
+7. Documentation and runbooks
+
+This phase will establish v1z3r as a production-ready, enterprise-grade application capable of handling significant scale while maintaining optimal performance.