… scripts/deploy-production.sh
… src/utils/alerting.ts
… src/utils/cachingStrategy.ts
… src/utils/connectionPool.ts
… src/utils/performanceMonitor.ts
… tests/load/load-test-config.js
… tests/load/run-load-tests.sh